Connect between transferring services
Connection is done by writing a custom groupBy method which is similar to a
regular groupBy with the addition of copying the last item of a group to the
first item of subsequent group. Both polylines and journey will make use of that
result. A bit messy, will definitely need to cleanup later